+++ src/PINdemonium/WxorXHandler.cpp
@@ -100,7 +100,7 @@ VOID WxorXHandler::_writeSetManager( ADDRINT start_addr, UINT32 size,std::vector
 	UINT32 end_addr = (UINT32)(start_addr + size);
 	//iterate through our structure in order to find if we have to update one of our WriteInterval
 	//We can't use an iterator because, after a certain amount of writeinterval, it will broke
-	for(int i = 0; i < currentWriteSet.size(); i++){
+	for(size_t i = 0; i < currentWriteSet.size(); i++){
 		//if we foud that an item has to be updated then update it and return
 		if(currentWriteSet[i].checkUpdate(start_addr, end_addr)){
 			currentWriteSet[i].update(start_addr, end_addr, isheap);
